requirements.txt
requests
pandas
numpy
scikit-learn
flask
kafka-python